requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
google-generativeai==0.7.0
